refactor(__init__.py): Pull out code for running a local server
